redesign the routed signals representation
Prior to this change, routed signals were represented by a dictionary where the
key was the source signal and the destination was in a sub-dictionary.  This
caused pretty obvious problems if a signal was routed to more than one
destination.  This change now uses a tuple of (source,destintation) as the key
to a sub-dictionary.

Some minor changes to help ni driver to shut down more cleanly, perhaps only
helpful in the simulated mode.