Subdevice Population in device.py
Changed subdevice names by adding a number to allow dictionaries to populate.
not sure if this will create bugs, but it should be useful infromation once
more ctypes functions are implemented.

Had to change the map given to clock/trig sources in subdevice.py,
but some devices still cant find a source. Will need to either fix routes
or add some logic here to give the right mapping to the correct subdevice.