Static Output
The static output function for comedi devices is working.
Currently the command data structure is used, but this will have to change
as waveform generation is developed. Routing is not implemented.
Static output should be possible to achieve from the GUI without any presets.

note: My verison of arbwave has no way to select device clocks from the GUI. This is not relevant for static output