Waveforms and Routing
This overdue update includes the implementation of anaolgue waveform generation
using a comedi command as well as a first pass at signal routing. Currently,
routing to PFI,RTSI are complete and analog output clocks and triggers are on
thier way. Digital waveforms are unimplemented as are GPCT commands. Waveform
output on multiple channels is broken. The only remaining component which
assumes an ni device is the clock search at the begining of the subdevice file.
Clocks do not populate properly in the gui.