Fix threading freezes.  Implement threading for gtk3
Threading interactions have changed in GTK at version > 3.10.  The current
version does not need to have so much manual labor.  On the other hand, if
one uses the old manual threads_init, threads_enter, and threads_leave,
one of the side effects is to freeze a window entirely under a thread
mutex lockup when the window is moved.  This change removes those problems.