Fixes: 1722d4ceb4 again; detect need for */SampleClock-->OnboardClock
This certainly works for the analog output devices where the SampleClock is
actually identified as "OnboardClock", with no prefixes, when it is used as
the SampleClock source.  Not sure when this change will break--if NI has
some other way of identifying some sort of an internal sample clock for
some device, but does not call it OnboardClock.