comedi: signals fixes; remove ctypes.byref;
Fixes lots of things with respect to signals so that sample clock is
correctly identified.  This represents the first bit of code that at least
does not complain when arbwave dictates some output on analog output
channels.

This also removes the use of ctypes.byref since we have updated the
supporting comedilib library to avoid segfaults and to properly auto-byref
arguments if they are the proper ctypes type.