re-implement distribution of finite_end_clock capability/requirement
This patch pushes the requirement specification for finite_end_clock lower
into the drivers.  This patch also defines and implements the rule for a
driver/channel that requires an end-clock in order to complete a
non-continuous output sequence:  The driver/channel must ignore the last
transition given to it if it instructs arbwave to generate a final clock.