comedi: implement new distribution of finite_end_clock
Re-implements new tactic for determining whether a channel requires a
end-clock in non-continuous mode.  This also implements, for comedi, the
rule that drivers/channels ignore the last transition if arbwave has been
instructed to generate a end-clock for that particular channel.

This patch adds a lookup table that is going to need to be kept up-to-date
with respect to drivers/boards--this is because comedi does not currently
provide the means to identify subdevices/channels that require an extra
clock at the end of a sequence when in finite-generation mode.