gx3500_dio: remove default clock; ext-trig not set by default
The default clock in gx3500_dio causes confusion as its clock _must_ be
identified to Arbwave as a clock.  Thus, it does not help at all to set a
default value, when one must assign this in any case.

Having external trigger set by default can be a problem if one just wants
to get things working without additional hardware configuration.

Note:  Something is wrong with the triggering in any case (in the
firmware).  There is also a mishandling of the triggering when in simulated
mode.