re-initialize locals for waveform generation
Local variables dictionary was earlier not re-initialized each time
waveform generation was begun.  This meant that old values stuck around in
the root-group local variables dictionary.  This was discovered when trying
to implement sympy expressions in Arbwave where expressions parameters
'expr_steps' and 'expr_err' were trying to be inherited from the global
environment but allowed to be overwritten at each new instance/copy of
the local environment.

This patch ensures that the root-group locals dictionary is created from
scratch for each new entry into the root-group waveform generator.