dds: drivers now return _all_ output channels together
The separate output channels (analog, digital) were never really treated
differently until waveform/output compute time (in compute.py) anyway.  In
fact, the different channels were generally grouped together.  As we add
new types of channels (DDS for the case of this branch), it seems better
to just lump them together.

So, this patch just lumps all the output channels together as the
information is queried from the driver to Arbwave.  The output values of
the different channel types are indeed still kept separate when passing
this information back to the drivers (this may change in the future).