gxt: change many debug lines to use old printf-style statements
This change is to improve speed of operation when debug statements are not
in effect.  Using the older printf-style way of formatting strings delays
the string formatting until it is required by the debug statement.  In the
case when the logging level is to high to allow debug statements to be
printed, the string formatting is not performed.