bbb: fix paths of clock/timing channels
The paths of the clock/timing channels needs to be full for external
interfaces, but not for those interfaces internal to the bbb drivers.  For
instance, the user sees a clock labeled as
bbb/{hostid}/timing/InternalClock, but the bbb driver only recognizes this
as the "InternalClock".  This change was to help keep the two of these
straight without also requiring a lot of conversions within the bbb driver.