bbb: only start() devs that have waveform data
Generally, Arbwave currently starts every device that is configured,
regardless of whether it has channels enabled or not.  It seems that
Arbwave ought to actually really skip the device if it has not clocks or
channel-waveform output configured.  This implements this at least for the
bbb devices.

Removes mistaken '*' operator -- caused syntax error.