Try to allow exact integer multiples for minimum_periods
If the minimum_period was created using an exact integer multiple of the
base clock period, it is very possible that the machine precision will
promote a float value to the next integer up.  This change ensures that any
minimum_period that is within (100*eps)% of an integer multiple of base
clocks is actualy promoted to this this exact multiple.

There is a possibility that this fix will have unintended side-effects for
other drivers...