Fix plotter zooming
When upgrading matplotlib and python, the zooming features mostly broke.
This was mostly due to comparisons in strings being performed with exact
'is' comparison instead of '==' equivalence comparison.

This change fixes those comparisons to use the '==' equivalence operator.
This patch also removes the canvas.widgetlock that was asserted during a
SpanSelector operation.  It appears that a newer version of matplotlib also
froze the SpanSelector widget when the lock was in place.