bbb: enable all possible digital channels
I guess I had only instantiated 4 channels as preliminary code--this patch
upgrades the total number channels to 10 or the maximum number of channels
possible.

The two last bits available are actually bits 14:15 in the backend bbb
driver/firmware.  This Arbwave backend driver for bbb maps channels 8:9 to
bits 14:15 for the hardware.