fix script editor cursor placement after save/apply
Changes in gtk3 vs gtk2 (or perhaps from Python->Python3?) caused a crash
in the entire gui often when saves were happening.  This had to do with the
iterator for a cursor position not being valid after an update to the text
buffer.  This patch fixes this by simply not updating the buffer when the
changes are from the editor back to the target storage.  In this way, we no
longer have to do any changes to the cursor or vertical position of the
buffer.  We do have to grab the focus back though.