Use info from range object to properly set adjustment info in int controls
Previous implementation used max() and min() to determine what the extent
of a range is.  This is a particularly problematic approach because it is a
very large computational load to find the maximum of an interable as large
as range(0, sys.maxsize).

(Besides, there is no reason to not use this information as it already
available.)