backend.service: provide means to listen on 0.0.0.0
Pyro4 does not normally support listening on 0.0.0.0, but this patch
provides a means to ensure that all URIs sent back to the client machine
include the ip address and port visible to the client.  Note that this
information does not necessarily match the actual address and port listened
to on the service machine (any number of tunnels may have obfuscated the
actual link).

The user can also specify the actual bind address using a command line
parameter to --service.