Fix spurious lines back to origin that were added by plotter
In non-continuous mode, extra data points are added to the data that is
intended on having the data end the experiment on a known configuration.

The problem was that the plotter was sorting the data points to plot by the
group label.  The group label should only be used as a unique identifier of
waveform elements, and not impose any meaning that this identifier has any
type of ordering/sorting.

This change also fixes the gnuplot export tool.