bbb:  pass clocks to set_waveforms for devices
Earlier, I had assumed that the clocks could be ignored for devices such as
DDS because this just needed to create its own output and separately let
the clocks perform their own work.  This is only true *if* it OK for the
DDS to be the only client to its clock source (and if the DDS clock source
is indeed aperiodic).  This change allows for the DDS in the future to
1) either complain that it is apparently sharing a clock with another
   device (less preferable)
or
2) insert null operations to pad the DDS waveform such that it can handle
   the extra clocks used by another device.

This patch also includes a minor log message change.