bbb:  improve timing/dds sim and complain of shared clocks for dds
This patch now passes the transitions of its clock to the dds controller.
This patch then also changes bbb.dds controller to complain when it detects
that another device has caused other clocks to be added to the clock source
(or because the base clock was periodic instead of aperiodic).

By splitting off many of the details of the bbb.timing/dds devices from
the controller into a separate file, these can now be incorporated into the
simulator to help with testing.