bbb analog:  Order waveform to allow use of LMBD instruction
The LMBD instruction used by the analog-lmbd branch on the bbb analog
device allows for a much faster iteration through all channels that
participate in a waveform.  The caveat is that the channel data must be
sorted from highest channel number to the lowest channel number.  This
patch ensures that channel data is ordered in most-significant-channel
order.