bbb analog: make spans auto-selected if possible
This changes the configuration so that the user is supposed to identify the
setting of the MSP[2:0] pins, rather than the span specifically.  The
actual spans used then now depends on whether MSP[2:0] is set to the
SoftSpan setting and depends on the scaling of the specific analog channel.

If MSP[2:0] pins do not allow for programmabiilty, the span for all
channels is set to the fixed MSP[2:0] setting.